requirements.txt
ddt==1.7.2
pytest==7.1.2
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
selenium==3.141.0
